+++ core/src/main/scala/cats/tagless/implicits/package.scala
@@ -16,4 +16,4 @@
 
 package cats.tagless
 
-package object implicits extends instances.AllInstances with syntax.AllSyntax
+package object implicits extends syntax.AllSyntax
