--- core/src/main/scala/cats/tagless/instances/AllInstances.scala
@@ -1,34 +1,0 @@
-/*
- * Copyright 2019 cats-tagless maintainers
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package cats.tagless.instances
-
-trait AllInstances extends EitherKInstances
-  with EitherTInstances
-  with FuncInstances
-  with IdTInstances
-  with IorTInstances
-  with KleisliInstances
-  with CokleisliInstances
-  with NestedInstances
-  with OneAndInstances
-  with OptionTInstances
-  with Tuple2KInstances
-  with WriterTInstances
-  with ContravariantKInstances
-  with InvariantKInstances
-  with FunctorKInstances
-  with SemigroupalKInstances
