+++ tests/src/test/scala/cats/tagless/tests/CatsDataInstancesTests.scala
@@ -29,19 +29,19 @@ import scala.util.Try
 class CatsDataInstancesTests extends CatsTaglessTestSuite {
   implicit def idKisomorphismK[A]: IsomorphismsK[IdK[A]#λ] = IsomorphismsK.invariantK[IdK[A]#λ]
 
-  checkAll("FunctorK[Nested[?[_], Eval, Int]]", FunctorKTests[Nested[?[_], Eval, Int]].functorK[Try, Option, List, Int])
-  checkAll("FunctorK[OneAnd[?[_], Boolean]]", FunctorKTests[OneAnd[?[_], Boolean]].functorK[Try, Option, List, Int])
-  checkAll("FunctorK[Tuple2K[Eval, ?[_], Int]]", FunctorKTests[Tuple2K[Eval, ?[_], Int]].functorK[Try, Option, List, Int])
-  checkAll("ContravariantK[Cokleisli[?[_], String, Int]]", ContravariantKTests[Cokleisli[?[_], String, Int]].contravariantK[Try, Option, List, Int])
+  checkAll("FunctorK[Nested[*[_], Eval, Int]]", FunctorKTests[Nested[*[_], Eval, Int]].functorK[Try, Option, List, Int])
+  checkAll("FunctorK[OneAnd[*[_], Boolean]]", FunctorKTests[OneAnd[*[_], Boolean]].functorK[Try, Option, List, Int])
+  checkAll("FunctorK[Tuple2K[Eval, *[_], Int]]", FunctorKTests[Tuple2K[Eval, *[_], Int]].functorK[Try, Option, List, Int])
+  checkAll("ContravariantK[Cokleisli[*[_], String, Int]]", ContravariantKTests[Cokleisli[*[_], String, Int]].contravariantK[Try, Option, List, Int])
   checkAll("ApplyK[IdK[Int]#λ]", ApplyKTests[IdK[Int]#λ].applyK[Try, Option, List, Int])
-  checkAll("ApplyK[EitherK[Eval, ?[_], Int]]", ApplyKTests[EitherK[Eval, ?[_], Int]].applyK[Try, Option, List, Int])
-  checkAll("ApplyK[EitherT[?[_], String, Int]]", ApplyKTests[EitherT[?[_], String, Int]].applyK[Try, Option, List, Int])
-  checkAll("ApplyK[Func[?[_], String, Int]]", ApplyKTests[Func[?[_], String, Int]].applyK[Try, Option, List, Int])
-  checkAll("ApplyK[IdT[?[_], Int]]", ApplyKTests[IdT[?[_], Int]].applyK[Try, Option, List, Int])
-  checkAll("ApplyK[IorT[?[_], String, Int]]", ApplyKTests[IorT[?[_], String, Int]].applyK[Try, Option, List, Int])
-  checkAll("ApplyK[Kleisli[?[_], String, Int]]", ApplyKTests[Kleisli[?[_], String, Int]].applyK[Try, Option, List, Int])
-  checkAll("ApplyK[OneAnd[?[_], Int]]", ApplyKTests[OneAnd[?[_], Int]].applyK[Try, Option, List, Int])
-  checkAll("ApplyK[OptionT[?[_], Int]]", ApplyKTests[OptionT[?[_], Int]].applyK[Try, Option, List, Int])
-  checkAll("ApplyK[Tuple2K[Set, ?[_], Int]]", ApplyKTests[Tuple2K[Set, ?[_], Int]].applyK[Try, Option, List, Int])
-  checkAll("ApplyK[WriterT[?[_], String, Int]]", ApplyKTests[WriterT[?[_], String, Int]].applyK[Try, Option, List, Int])
+  checkAll("ApplyK[EitherK[Eval, *[_], Int]]", ApplyKTests[EitherK[Eval, *[_], Int]].applyK[Try, Option, List, Int])
+  checkAll("ApplyK[EitherT[*[_], String, Int]]", ApplyKTests[EitherT[*[_], String, Int]].applyK[Try, Option, List, Int])
+  checkAll("ApplyK[Func[*[_], String, Int]]", ApplyKTests[Func[*[_], String, Int]].applyK[Try, Option, List, Int])
+  checkAll("ApplyK[IdT[*[_], Int]]", ApplyKTests[IdT[*[_], Int]].applyK[Try, Option, List, Int])
+  checkAll("ApplyK[IorT[*[_], String, Int]]", ApplyKTests[IorT[*[_], String, Int]].applyK[Try, Option, List, Int])
+  checkAll("ApplyK[Kleisli[*[_], String, Int]]", ApplyKTests[Kleisli[*[_], String, Int]].applyK[Try, Option, List, Int])
+  checkAll("ApplyK[OneAnd[*[_], Int]]", ApplyKTests[OneAnd[*[_], Int]].applyK[Try, Option, List, Int])
+  checkAll("ApplyK[OptionT[*[_], Int]]", ApplyKTests[OptionT[*[_], Int]].applyK[Try, Option, List, Int])
+  checkAll("ApplyK[Tuple2K[Set, *[_], Int]]", ApplyKTests[Tuple2K[Set, *[_], Int]].applyK[Try, Option, List, Int])
+  checkAll("ApplyK[WriterT[*[_], String, Int]]", ApplyKTests[WriterT[*[_], String, Int]].applyK[Try, Option, List, Int])
 }
