+++ tests/src/test/scala/cats/tagless/tests/autoApplyKTests.scala
@@ -23,7 +23,6 @@ import cats.instances.all._
 import cats.laws.discipline.SerializableTests
 import cats.laws.discipline.arbitrary._
 import cats.laws.discipline.eq._
-import cats.tagless.instances.all._
 import cats.tagless.laws.discipline.ApplyKTests
 import cats.tagless.tests.autoApplyKTests.AutoApplyKTestAlg
 import org.scalacheck.Arbitrary
